+++ src/gui/SetorNovoFormController.java
@@ -1,8 +1,10 @@
 package gui;
 
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
+import db.DbIntegrityException;
 import gui.util.Alerts;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -11,6 +13,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -22,9 +25,14 @@ public class SetorNovoFormController implements Initializable {
 
 	private Setor setor;
 
+	private Setor setorComparar;
+
 	private SetorService service;
-	
+
 	private static ObservableList<Setor> listaSetor;
+
+	@FXML
+	private Button btNovo;
 
 	@FXML
 	private Button btSalvar;
@@ -48,22 +56,74 @@ public class SetorNovoFormController implements Initializable {
 	private TableColumn<Setor, String> tableColumnNome;
 
 	@FXML
-	public void onBtSalvarSetorAction(ActionEvent event) {
+	public void onBtNovoAction(ActionEvent event) {
+
+		txtNome.setText("");
+		txtId.setText("");
+		txtNome.requestFocus();
+
+		Setor setorNovo = new Setor(null, "");
+
+		setSetor(setorNovo);
+
+	}
+
+	@FXML
+	public void onBtSalvarAction(ActionEvent event) {
 
 		setSetor(getFormData());
 
-		if (setor != null) {
+		if (this.setor != null) {
 
-			service.setorNovoOuEditar(setor);
+			boolean ok = false;
+
+			ok = compararCampos();
+
+			if (ok == false) {
+
+				service.setorNovoOuEditar(this.setor);
+				limparCampos();
+				updateTableView();
+
+			} else {
+
+				Alerts.showAlert("Setor", "Salvar / Editar", "Não houve alteração no registro", AlertType.INFORMATION);
+
+			}
 
 		}
 
 	}
-	
-	@FXML
-	public void onBtExcluirSetorAction(ActionEvent event) {
 
-		service.remove(setor);
+	@FXML
+	public void onBtExcluirAction(ActionEvent event) {
+
+		if (service == null) {
+			throw new IllegalThreadStateException("Service está nulo");
+		}
+		try {
+
+			if (setor.getIdSetor() != null) {
+
+				Optional<ButtonType> result = Alerts.showConfirmation("Confirmação", "Você quer deletar o setor?");
+
+				if (result.get() == ButtonType.OK) {
+
+					service.remove(setor);
+					updateTableView();
+
+				}
+
+			} else {
+
+				Alerts.showAlert("Setor", "Excluir", "Selecione um registro", AlertType.INFORMATION);
+
+			}
+
+		} catch (DbIntegrityException e) {
+			Alerts.showAlert("Setor", "Excluir", "Erro ao excluir o setor", AlertType.INFORMATION);
+			limparCampos();
+		}
 
 	}
 
@@ -83,25 +143,33 @@ public class SetorNovoFormController implements Initializable {
 	}
 
 	private void initializeNodes() {
-		
+		showDetails(null);
+
+		tableViewSetor.getSelectionModel().selectedItemProperty()
+				.addListener((observable, oldValue, newValue) -> showDetails(newValue));
+
 		tableColumnId.setCellValueFactory(new PropertyValueFactory<>("idSetor"));
 		tableColumnNome.setCellValueFactory(new PropertyValueFactory<>("Nome"));
-		
+
+		// Stage stage = (Stage) Main.getMainScene().getWindow();
+		// tableViewSetor.prefHeightProperty().bind(stage.heightProperty());
+
 		service = new SetorService();
 		setor = new Setor();
-		
+
+		updateTableView();
 	}
-	
+
 	public void updateTableView() {
+
 		if (service == null) {
-			throw new IllegalStateException("Service está nulo");
+			throw new IllegalStateException("Service nulo");
 		}
 
 		listaSetor = FXCollections.observableArrayList(service.findAll());
 		tableViewSetor.setItems(listaSetor);
-		
-	}
 
+	}
 
 	private Setor getFormData() {
 
@@ -116,8 +184,17 @@ public class SetorNovoFormController implements Initializable {
 			setor = null;
 
 		} else {
-			
-			setor.setIdSetor(Integer.valueOf(txtId.getText()));
+
+			if (txtId.getText().equals("")) {
+
+				setor.setIdSetor(null);
+
+			} else {
+
+				setor.setIdSetor(Integer.valueOf(txtId.getText()));
+
+			}
+
 			setor.setNome(txtNome.getText());
 
 		}
@@ -125,5 +202,48 @@ public class SetorNovoFormController implements Initializable {
 		return setor;
 
 	}
+
+	public void showDetails(Setor setor) {
+
+		if (setor != null) {
+
+			txtId.setText(String.valueOf(setor.getIdSetor()));
+			txtNome.setText(setor.getNome());
+			setSetor(setor);
+			setorComparar = this.setor;
+
+		} else {
+
+			limparCampos();
+
+		}
+
+	}
+
+	public void limparCampos() {
+		txtId.setText("");
+		txtNome.setText("");
+	};
+
+	public boolean compararCampos() {
+
+		boolean ok = false;
+
+		if (setorComparar == null) {
+
+			return ok;
+
+		} else if (this.setor.getNome().equals(setorComparar.getNome())) {
+
+			ok = true;
+			return ok;
+
+		} else {
+
+			return ok;
+
+		}
+
+	};
 
 }
